+++ test_server.sh
@@ -1,2 +1,2 @@
 # Run testing AioHTTP app on the given port
-PYTHONPATH=$(pwd)/src python test/app.py "$@"
+PYTHONPATH=$(pwd)/src python test/app/__init__.py "$@"